+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/handler/ControllerScanner.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.servlet.mvc.handler;
 
 import com.interface21.context.stereotype.Controller;
 import org.reflections.Reflections;
