+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/view/ModelAndViewConverter.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.servlet.mvc.view;
 
 import com.interface21.webmvc.servlet.ModelAndView;
