+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/view/ViewResolver.java
@@ -1,8 +1,6 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.servlet.mvc.view;
 
 import com.interface21.webmvc.servlet.ModelAndView;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 
 public interface ViewResolver {
